pyproject.toml
[project]
name = "recursive-language-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]
